requirements.txt
absl-py==2.1.0
anyio==4.3.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.2.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Eel==0.16.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
fastjsonschema==2.19.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
future==1.0.0
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
graphics==1.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
ipykernel==6.19.2
ipython==8.7.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.2
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.26.0
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
mando==0.7.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
namex==0.0.7
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.5.6
nltk==3.8.1
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==22.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==2.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.36
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
Pygments==2.13.0
pyparsing==3.1.2
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==24.0.1
radon==6.0.1
referencing==0.34.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
Send2Trash==1.8.3
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soupsieve==2.5
stack-data==0.6.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.66.4
traitlets==5.7.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
wavio==0.0.8
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.2
whichcraft==0.6.1
wrapt==1.16.0
zope.event==5.0
zope.interface==6.3
